+++ Medinilla.DataTypes/Core/TransactionConsumption.cs
@@ -6,7 +6,7 @@ public sealed class TransactionConsumption
 {
     public ConsumptionType ConsumptionType { get; set; }
 
-    public decimal Consumption { get; set; }
+    public float Consumption { get; set; }
 
     public DateTime Timestamp { get; set; }
 }
